Intorudce abstract converter for section containers.
